Refactor

+++ src/evaluation/eval_data_obj.py
@@ -10,12 +10,12 @@ class EvalDataObj():
     def load_val_data(self):
         states_val_arr = np.loadtxt('./src/datasets/states_val_arr.csv', delimiter=',')
         targets_val_arr = np.loadtxt('./src/datasets/targets_val_arr.csv', delimiter=',')
-        return states_val_arr, targets_val_arr
+        return np.float32(states_val_arr), np.float32(targets_val_arr)
 
     def load_test_episode_ids(self, traffic_density):
         assert traffic_density == '' or \
                                         traffic_density == 'high_density' or \
-                                        traffic_density == 'medium_density'  
+                                        traffic_density == 'medium_density'
 
         if traffic_density == '':
             dir_path = f'./src/datasets/validation_episodes.csv'
